fix(vincent-rest-api): unused imports, unneeded files
