+++ packages/vincent-rest-api/package.json
@@ -4,11 +4,12 @@
   "description": "Integration tests for Vincent Registry API",
   "packageManager": "pnpm@10.7.0",
   "scripts": {
-    "generate-openapi": "ts-node src/generate-openapi.ts",
-    "generate-rtk-client": "npx @rtk-query/codegen-openapi openapi-config.js",
+    "generate-openapi": "pnpx ts-node src/utils/generate-openapi.ts",
+    "generate-rtk-client": "pnpx @rtk-query/codegen-openapi openapi-config.js",
     "generate-api": "run-s generate-openapi generate-rtk-client"
   },
   "devDependencies": {
+    "@asteasolutions/zod-to-openapi": "^5.5.0",
     "@reduxjs/toolkit": "^2.8.2",
     "@rtk-query/codegen-openapi": "^1.2.0",
     "@types/express": "^4.17.21",
@@ -18,7 +19,6 @@
     "eslint": "^8.0.0",
     "jest": "^29.5.0",
     "npm-run-all": "^4.1.5",
-    "openapi-typescript-codegen": "^0.25.0",
     "react-redux": "^9.2.0",
     "ts-jest": "^29.1.0",
     "ts-node": "^10.9.2",
@@ -26,9 +26,7 @@
     "unbuild": "^1.0.0"
   },
   "dependencies": {
-    "@asteasolutions/zod-to-openapi": "^5.5.0",
     "express": "^4.18.2",
-    "js-yaml": "^4.1.0",
     "zod": "^3.22.2"
   }
 }
